Rebase with dev and move some of the refactoring to a separate branch
