+++ packages/sitecore-jss/src/data-api.ts
@@ -1,6 +1,5 @@
 import { LayoutServiceData, PlaceholderData } from './data-models';
 import { HttpDataFetcher, HttpResponse } from './data-fetcher';
-import { getQueryString } from './util/routing-utils';
 
 class ResponseError extends Error {
   response: HttpResponse<unknown>;
@@ -23,6 +22,18 @@ function checkStatus<T>(response: HttpResponse<T>) {
 
   const error = new ResponseError(response.statusText, response);
   throw error;
+}
+
+// note: encodeURIComponent is available via browser (window) or natively in node.js
+// if you use another js engine for server-side rendering you may not have native encodeURIComponent
+// and would then need to install a package for that functionality
+/**
+ * @param {Object} params
+ */
+function getQueryString(params: { [key: string]: string | number | boolean }): string {
+  return Object.keys(params)
+    .map((k) => `${encodeURIComponent(k)}=${encodeURIComponent(params[k])}`)
+    .join('&');
 }
 
 /**
